# Conflicts:
#	app/src/main/java/org/cxct/sportlottery/network/common/MatchType.kt
#	app/src/main/java/org/cxct/sportlottery/ui/game/GameViewModel.kt
#	app/src/main/java/org/cxct/sportlottery/ui/game/hall/GameV3Fragment.kt

+++ app/src/main/java/org/cxct/sportlottery/ui/bet/list/BetButton.kt
@@ -62,7 +62,7 @@ class BetButton @JvmOverloads constructor(
 
     private fun setupQuotaListener() {
         tv_quota.doAfterTextChanged {
-            isCanSendOut = it.toString().replace(",", "").toDouble() != 0.0
+            isCanSendOut = it.toString().replace(",", "").toInt() != 0
         }
     }
